+++ tests/inline.test.ts
@@ -1,5 +1,6 @@
 import { assertSiteSnapshot, build, getSite } from "./utils.ts";
 import inline from "../plugins/inline.ts";
+import basePath from "../plugins/base_path.ts";
 
 Deno.test("inline plugin", async (t) => {
   const site = getSite({
@@ -26,6 +27,22 @@ Deno.test("inline plugin (sourceURL)", async (t) => {
     copyAttributes: ["custom", /^data-/, /^@/],
     sourceURL: true,
   }));
+
+  site.add([".svg", ".js", ".png"]);
+  site.add("favicon.png", "favicon2.png");
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});
+
+Deno.test("inline plugin (basePath)", async (t) => {
+  const site = getSite({
+    location: new URL("https://example.com/blog/"),
+    src: "inline",
+  });
+
+  site.use(basePath());
+  site.use(inline());
 
   site.add([".svg", ".js", ".png"]);
   site.add("favicon.png", "favicon2.png");