+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - Ignore `deno.json` and `deno.jsonc` files by the watcher.
 
 ### Fixed
+- `inline` plugin: Error when basePath plugin is used before.
 - `icons` plugin: clear the plugin list on update.
 - Updated dependencies: `vento`, `satori`, `sharp`, `esbuild`, `deno-loader`, `sass` and some icons.